adds BaseButton file + FakeButton option
