+++ src/components/ButtonSystem/BaseButton/index.ts
@@ -1,0 +1,3 @@
+import BaseButton from './BaseButton';
+
+export default BaseButton;
